Removed unnecessary Short.valueOf()
git-svn-id: https://ha-jdbc.svn.sourceforge.net/svnroot/ha-jdbc/trunk@784 882a69a1-a10d-0410-8702-ae3607fe02de

+++ ha-jdbc/src/net/sf/hajdbc/sync/DifferentialSynchronizationStrategy.java
@@ -123,7 +123,7 @@ public class DifferentialSynchronizationStrategy extends AbstractSynchronization
 						String name = primaryKeyResultSet.getString("COLUMN_NAME");
 						short position = primaryKeyResultSet.getShort("KEY_SEQ");
 		
-						primaryKeyColumnMap.put(Short.valueOf(position), name);
+						primaryKeyColumnMap.put(position, name);
 						
 						primaryKeyName = primaryKeyResultSet.getString("PK_NAME");
 					}
